requirements.txt
anyio==4.6.2.post1
APScheduler==4.0.0a5
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.5
packaging==24.2
pika==1.3.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
waitress==3.0.1
Werkzeug==3.1.2
WTForms==3.2.1
